test: share notice-test stubs and the workspace fingerprint
The no-op Git/Content/Editor ports and the byte-level workspace
fingerprint were declared three times across the remote-notice suites;
they now live in tests/common. Instrumented spies stay local to the
safety suite, which is their point.
