+++ tests/remote_notice_startup.rs
@@ -2,13 +2,10 @@
 
 mod common;
 
-use common::TempDir;
+use common::{NoopContent, NoopEditor, NoopGit, TempDir};
 use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
-use herdr_file_viewer::controller::{
-    Components, ContentProvider, Controller, EditorHandoff, EditorOutcome, GitService,
-    RenderResult, RootProviders,
-};
-use herdr_file_viewer::git::{Baseline, Status};
+use herdr_file_viewer::controller::{Components, Controller, RootProviders};
+use herdr_file_viewer::git::Baseline;
 use herdr_file_viewer::intent::Intent;
 use herdr_file_viewer::presenter;
 use herdr_file_viewer::update::cache::Cache;
@@ -17,55 +14,12 @@ use herdr_file_viewer::update::{
     DiscoveryRunner, NoticeSnapshot, ObjectId, ReleaseState, ReleaseTag, Source, StartDeps,
     UpdateState, Version, start_with,
 };
-use herdr_file_viewer::view_policy::ViewMode;
 use ratatui::Terminal;
 use ratatui::backend::TestBackend;
-use ratatui::text::Text;
-use std::collections::BTreeMap;
 use std::path::Path;
 use std::sync::Arc;
 use std::sync::mpsc;
 use std::time::Duration;
-
-struct Git;
-
-impl GitService for Git {
-    fn status(&self) -> BTreeMap<std::path::PathBuf, Status> {
-        BTreeMap::new()
-    }
-
-    fn changed_set(&self, _baseline: Baseline) -> BTreeMap<std::path::PathBuf, Status> {
-        BTreeMap::new()
-    }
-
-    fn diff(&self, _path: &Path, _baseline: Baseline, _full_context: bool) -> String {
-        String::new()
-    }
-
-    fn diff_directory(&self, _path: &Path, _baseline: Baseline) -> String {
-        String::new()
-    }
-}
-
-struct Content;
-
-impl ContentProvider for Content {
-    fn render(&self, _path: &Path, _mode: ViewMode, _diff: Option<&str>) -> RenderResult {
-        RenderResult {
-            content: Text::raw("test content"),
-            notices: Vec::new(),
-            source: None,
-        }
-    }
-}
-
-struct Editor;
-
-impl EditorHandoff for Editor {
-    fn open(&mut self, _file: &Path) -> EditorOutcome {
-        EditorOutcome::NoTakeover
-    }
-}
 
 #[derive(Clone)]
 struct Documents {
@@ -97,10 +51,10 @@ fn controller(dir: &Path) -> Controller {
         Baseline::Head,
         Components {
             providers: Box::new(|_| RootProviders {
-                git: Arc::new(Git),
-                content: Box::new(Content),
+                git: Arc::new(NoopGit),
+                content: Box::new(NoopContent),
             }),
-            editor: Box::new(Editor),
+            editor: Box::new(NoopEditor),
             clipboard: Box::new(common::RecordingClipboard::default()),
             renderers: None,
         },
